+++ src/main/java/io/github/xf8b/morefeatures/blocks/DisplayCaseBlock.java
@@ -1,0 +1,63 @@
+package io.github.xf8b.morefeatures.blocks;
+
+import io.github.xf8b.morefeatures.core.MoreFeaturesRegistries;
+import io.github.xf8b.morefeatures.tileentity.DisplayCaseTileEntity;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.SoundType;
+import net.minecraft.block.material.Material;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.ServerPlayerEntity;
+import net.minecraft.inventory.InventoryHelper;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.ActionResultType;
+import net.minecraft.util.Hand;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.BlockRayTraceResult;
+import net.minecraft.world.IBlockReader;
+import net.minecraft.world.World;
+import net.minecraftforge.fml.network.NetworkHooks;
+
+import javax.annotation.Nullable;
+
+public class DisplayCaseBlock extends Block {
+    public DisplayCaseBlock() {
+        super(Block.Properties.create(Material.GLASS)
+                .hardnessAndResistance(0.4f)
+                .sound(SoundType.GLASS)
+                .notSolid()
+        );
+    }
+
+    @Override
+    public boolean hasTileEntity(BlockState state) {
+        return true;
+    }
+
+    @Nullable
+    @Override
+    public TileEntity createTileEntity(BlockState state, IBlockReader world) {
+        return MoreFeaturesRegistries.DISPLAY_CASE_TILE_ENTITY.get().create();
+    }
+
+    @Override
+    public ActionResultType onBlockActivated(BlockState state, World worldIn, BlockPos pos, PlayerEntity player, Hand handIn, BlockRayTraceResult hit) {
+        if (!worldIn.isRemote) {
+            TileEntity tileEntity = worldIn.getTileEntity(pos);
+            if (tileEntity instanceof DisplayCaseTileEntity) {
+                NetworkHooks.openGui((ServerPlayerEntity) player, (DisplayCaseTileEntity) tileEntity, pos);
+            }
+        }
+        return ActionResultType.SUCCESS;
+    }
+
+    @Override
+    public void onReplaced(BlockState state, World worldIn, BlockPos pos, BlockState newState, boolean isMoving) {
+        if (state.getBlock() != newState.getBlock()) {
+            TileEntity tileEntity = worldIn.getTileEntity(pos);
+            if (tileEntity instanceof DisplayCaseTileEntity) {
+                InventoryHelper.dropItems(worldIn, pos, ((DisplayCaseTileEntity) tileEntity).getItems());
+            }
+        }
+    }
+}