+++ src/main/java/io/github/xf8b/morefeatures/blocks/IFoodDroppingLeaves.java
@@ -1,7 +1,10 @@
 package io.github.xf8b.morefeatures.blocks;
 
+import net.minecraft.block.Block;
 import net.minecraft.item.Item;
 
 public interface IFoodDroppingLeaves {
     Item getDroppedFood();
+
+    Block getSapling();
 }