+++ src/main/java/io/github/xf8b/morefeatures/blocks/ISaplingDroppingLeaves.java
@@ -1,7 +1,0 @@
-package io.github.xf8b.morefeatures.blocks;
-
-import net.minecraft.block.Block;
-
-public interface ISaplingDroppingLeaves {
-    Block getSapling();
-}