+++ src/main/java/io/github/xf8b/morefeatures/container/DisplayCaseContainer.java
@@ -1,0 +1,87 @@
+package io.github.xf8b.morefeatures.container;
+
+import io.github.xf8b.morefeatures.core.MoreFeaturesRegistries;
+import io.github.xf8b.morefeatures.tileentity.DisplayCaseTileEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.container.Container;
+import net.minecraft.inventory.container.Slot;
+import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketBuffer;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.IWorldPosCallable;
+
+import java.util.Objects;
+
+public class DisplayCaseContainer extends Container {
+    public final DisplayCaseTileEntity tileEntity;
+    private final IWorldPosCallable canInteractWithCallable;
+
+    public DisplayCaseContainer(final int windowId, final PlayerInventory playerInventory, final DisplayCaseTileEntity tileEntity) {
+        super(MoreFeaturesRegistries.DISPLAY_CASE_CONTAINER.get(), windowId);
+        this.tileEntity = tileEntity;
+        this.canInteractWithCallable = IWorldPosCallable.of(tileEntity.getWorld(), tileEntity.getPos());
+
+        int startX = 80;
+        int startY = 36;
+        int slotSizePlusTwo = 18;
+        this.addSlot(new Slot(tileEntity, 0, startX, startY));
+
+        int startPlayerInventoryX = 8;
+        int startPlayerInventoryY = 84;
+        for (int row = 0; row < 3; ++row) {
+            for (int column = 0; column < 9; ++column) {
+                this.addSlot(new Slot(playerInventory, 9 + (row * 9) + column, startPlayerInventoryX + (column * slotSizePlusTwo), startPlayerInventoryY + (row * slotSizePlusTwo)));
+            }
+        }
+
+        int hotbarY = 142;
+        for (int column = 0; column < 9; ++column) {
+            this.addSlot(new Slot(playerInventory, column, startPlayerInventoryX + (column * slotSizePlusTwo), hotbarY));
+        }
+    }
+
+    public DisplayCaseContainer(final int windowId, final PlayerInventory playerInv, final PacketBuffer data) {
+        this(windowId, playerInv, getTileEntity(playerInv, data));
+    }
+
+    private static DisplayCaseTileEntity getTileEntity(final PlayerInventory playerInv, final PacketBuffer data) {
+        Objects.requireNonNull(playerInv, "playerInv cannot be null");
+        Objects.requireNonNull(data, "data cannot be null");
+        final TileEntity tileAtPos = playerInv.player.world.getTileEntity(data.readBlockPos());
+        if (tileAtPos instanceof DisplayCaseTileEntity) {
+            return (DisplayCaseTileEntity) tileAtPos;
+        }
+        throw new IllegalStateException("Tile Entity is not correct! " + tileAtPos);
+    }
+
+    @Override
+    public boolean canInteractWith(PlayerEntity playerIn) {
+        return isWithinUsableDistance(canInteractWithCallable, playerIn, MoreFeaturesRegistries.DISPLAY_CASE.get());
+    }
+
+    @Override
+    public ItemStack transferStackInSlot(PlayerEntity playerIn, int index) {
+        ItemStack itemstack = ItemStack.EMPTY;
+        Slot slot = this.inventorySlots.get(index);
+        if (slot != null && slot.getHasStack()) {
+            ItemStack itemstack1 = slot.getStack();
+            itemstack = itemstack1.copy();
+            if (index < 1) {
+                if (!this.mergeItemStack(itemstack1, 1, this.inventorySlots.size(), true)) {
+                    return ItemStack.EMPTY;
+                }
+            } else if (!this.mergeItemStack(itemstack1, 0, 1, false)) {
+                return ItemStack.EMPTY;
+            }
+
+            if (itemstack1.isEmpty()) {
+                slot.putStack(ItemStack.EMPTY);
+            } else {
+                slot.onSlotChanged();
+            }
+        }
+
+        return itemstack;
+    }
+}